fix(fitted): merge Supabase auth into proxy.ts, delete middleware.ts
Next.js 16 does not allow both middleware.ts and proxy.ts. Moves
Supabase Auth session refresh logic for /fitted/* routes into the
existing proxy.ts alongside admin JWT auth and CSP headers.
