+++ apps/root/src/proxy.ts
@@ -1,5 +1,6 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { jwtVerify } from 'jose';
+import { createServerClient } from '@supabase/ssr';
 import {
   allowedOrigins,
   allowedImageOrigins,
@@ -32,20 +33,7 @@ async function isValidAdminSession(
   }
 }
 
-export async function proxy(request: NextRequest) {
-  if (request.nextUrl.pathname.startsWith('/tools/admin')) {
-    const token = request.cookies.get(ADMIN_SESSION_COOKIE)?.value;
-    if (!(await isValidAdminSession(token))) {
-      const loginUrl = new URL('/tools/login', request.url);
-      loginUrl.searchParams.set(
-        'next',
-        request.nextUrl.pathname + request.nextUrl.search
-      );
-      return NextResponse.redirect(loginUrl);
-    }
-  }
-
-  const nonce = Buffer.from(crypto.randomUUID()).toString('base64');
+function buildCspValue(request: NextRequest, nonce: string): string {
   const cspHeader = `
     default-src 'self';
     script-src 'self' 'nonce-${nonce}' 'strict-dynamic' https: ${
@@ -65,27 +53,132 @@ export async function proxy(request: NextRequest) {
     connect-src 'self' ${allowedOrigins.join(' ')};
     img-src 'self' blob: data: ${allowedImageOrigins.join(' ')};
 `;
-  const contentSecurityPolicyHeaderValue = cspHeader
-    .replace(/\s{2,}/g, ' ')
-    .trim();
+  return cspHeader.replace(/\s{2,}/g, ' ').trim();
+}
 
+/**
+ * Handles Supabase Auth session refresh for /fitted/* routes.
+ *
+ * Refreshes the auth session on every request (keeps cookies in sync),
+ * redirects unauthenticated users to /fitted/login, and applies CSP
+ * headers to the response.
+ */
+async function handleFittedAuth(
+  request: NextRequest,
+  nonce: string,
+  cspValue: string
+): Promise<NextResponse> {
+  const supabaseUrl = process.env['NEXT_PUBLIC_SUPABASE_URL'];
+  const anonKey = process.env['NEXT_PUBLIC_SUPABASE_ANON_ID'];
+
+  if (!supabaseUrl || !anonKey) {
+    // If Supabase env vars are missing, let the request through
+    // (will fail at the page level with a clearer error)
+    const requestHeaders = new Headers(request.headers);
+    requestHeaders.set('x-nonce', nonce);
+    requestHeaders.set('Content-Security-Policy', cspValue);
+    const response = NextResponse.next({
+      request: { headers: requestHeaders },
+    });
+    response.headers.set('Content-Security-Policy', cspValue);
+    return response;
+  }
+
+  // Set nonce on request headers so the app can read it via headers().
+  // request.headers is mutable in Next.js proxy context — Supabase's
+  // setAll uses request.cookies.set() which mutates headers the same way.
+  request.headers.set('x-nonce', nonce);
+  request.headers.set('Content-Security-Policy', cspValue);
+
+  let supabaseResponse = NextResponse.next({ request });
+
+  const supabase = createServerClient(supabaseUrl, anonKey, {
+    cookies: {
+      getAll() {
+        return request.cookies.getAll();
+      },
+      setAll(cookiesToSet) {
+        for (const { name, value } of cookiesToSet) {
+          request.cookies.set(name, value);
+        }
+        supabaseResponse = NextResponse.next({ request });
+        for (const { name, value, options } of cookiesToSet) {
+          supabaseResponse.cookies.set(name, value, options);
+        }
+      },
+    },
+  });
+
+  // IMPORTANT: Do not add code between createServerClient and auth.getUser().
+  // A simple mistake could make it very hard to debug issues with users being
+  // randomly logged out.
+  const {
+    data: { user },
+  } = await supabase.auth.getUser();
+
+  const { pathname } = request.nextUrl;
+
+  // Allow login and auth callback routes without authentication
+  if (
+    pathname.startsWith('/fitted/login') ||
+    pathname.startsWith('/fitted/auth')
+  ) {
+    // If user is already logged in and visiting login, redirect to /fitted
+    if (user && pathname.startsWith('/fitted/login')) {
+      const url = request.nextUrl.clone();
+      url.pathname = '/fitted';
+      return NextResponse.redirect(url);
+    }
+    supabaseResponse.headers.set('Content-Security-Policy', cspValue);
+    return supabaseResponse;
+  }
+
+  // Redirect unauthenticated users to login
+  if (!user) {
+    const url = request.nextUrl.clone();
+    url.pathname = '/fitted/login';
+    return NextResponse.redirect(url);
+  }
+
+  // IMPORTANT: Return the supabaseResponse object as-is (with CSP added)
+  // to keep browser and server cookies in sync.
+  supabaseResponse.headers.set('Content-Security-Policy', cspValue);
+  return supabaseResponse;
+}
+
+export async function proxy(request: NextRequest) {
+  // Admin JWT auth for /tools/admin
+  if (request.nextUrl.pathname.startsWith('/tools/admin')) {
+    const token = request.cookies.get(ADMIN_SESSION_COOKIE)?.value;
+    if (!(await isValidAdminSession(token))) {
+      const loginUrl = new URL('/tools/login', request.url);
+      loginUrl.searchParams.set(
+        'next',
+        request.nextUrl.pathname + request.nextUrl.search
+      );
+      return NextResponse.redirect(loginUrl);
+    }
+  }
+
+  const nonce = Buffer.from(crypto.randomUUID()).toString('base64');
+  const cspValue = buildCspValue(request, nonce);
+
+  // Supabase Auth for /fitted/* routes
+  if (request.nextUrl.pathname.startsWith('/fitted')) {
+    return handleFittedAuth(request, nonce, cspValue);
+  }
+
+  // Default: apply CSP headers to all other routes
   const requestHeaders = new Headers(request.headers);
   requestHeaders.set('x-nonce', nonce);
-
-  requestHeaders.set(
-    'Content-Security-Policy',
-    contentSecurityPolicyHeaderValue
-  );
+  requestHeaders.set('Content-Security-Policy', cspValue);
 
   const response = NextResponse.next({
     request: {
       headers: requestHeaders,
     },
   });
-  response.headers.set(
-    'Content-Security-Policy',
-    contentSecurityPolicyHeaderValue
-  );
+  response.headers.set('Content-Security-Policy', cspValue);
 
   return response;
 }
